+++ semcore/data-table/CHANGELOG.md
@@ -19,6 +19,10 @@ CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangel
 - `selectedRows` from indexes to `UNIQ_ROW_KEY` values.
 - Width of accordion is always `100%` instead of dynamic calculation.
 
+### Fixed
+
+- Table scrolls up a bit on every interaction in header.
+
 ## [16.0.4] - 2025-05-30
 
 ### Fixed